+++ reana_job_controller/slurmcern_job_manager.py
@@ -111,7 +111,7 @@ class SlurmJobManagerCERN(JobManager):
     @JobManager.execution_hook
     def execute(self):
         """Execute / submit a job with Slurm."""
-        self.cmd = self._encode_cmd(" ".join(self.cmd))
+        self.cmd = self._encode_cmd(self.cmd)
         initialize_krb5_token(workflow_uuid=self.workflow_uuid)
         self.slurm_connection = SSHClient(
             hostname=SlurmJobManagerCERN.SLURM_HEADNODE_HOSTNAME,